+++ my-scheme-compiler/tests/compiler_tests.rs
@@ -57,11 +57,29 @@ fn test_simple_compilation() {
         "#(0 1 2 3 4 5 6 7 8 9)\n",
     );
     test(
-        "(letrec ((f (lambda (x) (g x)))
-                   (g (lambda (x) (+ 1 x))))
-           (f 6))",
-        "7\n",
+        "(letrec ((is-even?
+                    (lambda (x)
+                      (if (= x 0)
+                        #t
+                        (is-odd? (- x 1)))))
+                  (is-odd?
+                    (lambda (x)
+                      (if (= x 0)
+                        #f
+                        (is-even? (- x 1))))))
+           (is-even? 1000))",
+        "#t\n",
     );
+    // NOTE without the guarantee of tco the previous example can stack overflow
+    // in testing at 10,000
+    // Obviously, this makes using scheme a real bummer without a lazy sequence
+    // or using do everywhere we would blow the stack quite easily with letrec
+    // for even simple functions. Possibly on full optimization this might be
+    // optimized, but without the guarantee it is not exactly great to compile
+    // to rust. Would need to do like closure and decompose all recursion into
+    // do (like loop/recur) or build in a lazy sequence to allow for stackless
+    // recursion.
+
     /*
     test("(cons 1 2)", "(1 . 2)\n");
     test(
