+++ my-scheme-compiler/tests/test_program/src/main.rs
@@ -1,9 +1,50 @@
-use my_scheme_lib::data::value::Value;
 use my_scheme_lib::data::err::Error;
-use my_scheme_lib::proc::help::*;
+use my_scheme_lib::data::value::Value;
 use my_scheme_lib::list;
+use my_scheme_lib::proc::help::*;
 pub fn module(env: Environment) -> Result<Value, Error> {
-{ let env = push(&env); put(&env, "f", Value::Undefined); put(&env, "g", Value::Undefined); set(&env, "f", lambda(None, env.clone(), |args, env| { put(&env, "x", args[0].clone()); apply( get(&env, "g")?, vec![get(&env, "x")?] ) })?)?; set(&env, "g", lambda(None, env.clone(), |args, env| { put(&env, "x", args[0].clone()); apply( get(&env, "+")?, vec![Value::from(1), get(&env, "x")?] ) })?)?; apply( get(&env, "f")?, vec![Value::from(6)] ) }
+    {
+        let env = push(&env);
+        put(&env, "is-even?", Value::Undefined);
+        put(&env, "is-odd?", Value::Undefined);
+        set(
+            &env,
+            "is-even?",
+            lambda(None, env.clone(), |args, env| {
+                put(&env, "x", args[0].clone());
+                if apply(get(&env, "=")?, vec![get(&env, "x")?, Value::from(0)])?.is_true() {
+                    Ok(Value::Bool(true))
+                } else {
+                    apply(
+                        get(&env, "is-odd?")?,
+                        vec![apply(
+                            get(&env, "-")?,
+                            vec![get(&env, "x")?, Value::from(1)],
+                        )?],
+                    )
+                }
+            })?,
+        )?;
+        set(
+            &env,
+            "is-odd?",
+            lambda(None, env.clone(), |args, env| {
+                put(&env, "x", args[0].clone());
+                if apply(get(&env, "=")?, vec![get(&env, "x")?, Value::from(0)])?.is_true() {
+                    Ok(Value::Bool(false))
+                } else {
+                    apply(
+                        get(&env, "is-even?")?,
+                        vec![apply(
+                            get(&env, "-")?,
+                            vec![get(&env, "x")?, Value::from(1)],
+                        )?],
+                    )
+                }
+            })?,
+        )?;
+        apply(get(&env, "is-even?")?, vec![Value::from(10000)])
+    }
 }
 
 fn main() {
@@ -11,5 +52,5 @@ fn main() {
     match module(env) {
         Ok(val) => println!("{val}"),
         Err(e) => println!("{:?}", e),
-   }
+    }
 }
