--- my-scheme-lib/src/compile.rs
@@ -1,3 +1,0 @@
-pub mod compiler;
-pub mod core;
-pub mod proc;
