+++ my-scheme-lib/src/proc.rs
@@ -1,6 +1,7 @@
 pub mod arrays;
 pub mod chars;
 pub mod env;
+pub mod help;
 pub mod lists;
 pub mod numbers;
 pub mod others;
