+++ my-scheme-lib/tests/standard_6_3_5_strings.rs
