--- my-scheme-lib/tests/test_program/.gitignore
@@ -1,3 +1,0 @@
-/target
-/Cargo.lock
-/src/scheme_libs
