--- my-scheme-lib/tests/test_program/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "test-program"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-my-scheme-lib = { path="../../" }
